Refactor: Calendar모달에서 캐러셀 분리
