+++ FE/src/components/CalendarCourousel/Calendar/style.js
