+++ FE/src/components/CalendarCourousel/style.js
